Extract DebugMenu.java
